+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This is the Changelog for the Visual Studio Code Simple HTTP Server.
 
+## [2.0.0]
+- Fixed scroll level resetting on refresh.
+- Added command "Simple HTTP Server: Create HTTP Server With Current File" to create a HTTP Server with a single file!
+- Added better error handling.
+- Added more descreptions on how to use the features.
+
 ## [1.0.3]
 - Removed unneeded code
 - Added default settings css to the body for all Websites